requirements.txt
opencv==4.9.0
face_recognition==1.2.3
tkcalendar==1.5.0
mysql.connector==8.2.0